+++ CHANGELOG.md
@@ -1,3 +1,79 @@
+## [2.16.2](https://github.com/themadorg/madmail/compare/v2.16.1...v2.16.2) (2026-07-18)
+
+
+### Bug Fixes
+
+* **www:** correct UTF-8 corruption in Go-template-to-Minijinja conversion ([1bf1759](https://github.com/themadorg/madmail/commit/1bf17598f052cf2d8218260528f3c8150ae38a05))
+
+## [2.16.1](https://github.com/themadorg/madmail/compare/v2.16.0...v2.16.1) (2026-07-16)
+
+
+### Bug Fixes
+
+* **webimap:** ship multi-mailbox search/create via green CI ([5a31600](https://github.com/themadorg/madmail/commit/5a316002f6d8e81937fba6dea1993f512229fbe6))
+
+# [2.16.0](https://github.com/themadorg/madmail/compare/v2.15.0...v2.16.0) (2026-07-15)
+
+
+### Bug Fixes
+
+* **www:** convert Go if-not template actions for custom www_dir ([93d350d](https://github.com/themadorg/madmail/commit/93d350d074e6437bc3af584486ef935043736e3d))
+* **www:** silence clippy on html-migrate walk/tests ([57a6301](https://github.com/themadorg/madmail/commit/57a63014e828ad4667385234cb487e159a41212d))
+
+
+### Features
+
+* **www:** migrate custom Go templates on update ([c857e23](https://github.com/themadorg/madmail/commit/c857e2380ccbdac046bf3459e8f4688e5a9e3142))
+
+# [2.15.0](https://github.com/themadorg/madmail/compare/v2.14.0...v2.15.0) (2026-07-15)
+
+
+### Features
+
+* **logging:** flexible bool flags and maddy-compatible log destinations ([4c45778](https://github.com/themadorg/madmail/commit/4c45778ea6982013ea469e1370347b6dafd184f7))
+
+# [2.14.0](https://github.com/themadorg/madmail/compare/v2.13.5...v2.14.0) (2026-07-13)
+
+
+### Features
+
+* **webimap:** implement multi-mailbox WS actions for madcore parity ([bc58b0c](https://github.com/themadorg/madmail/commit/bc58b0c63f2c4611c3077edff2fedad6adbc4dd2))
+
+## [2.13.5](https://github.com/themadorg/madmail/compare/v2.13.4...v2.13.5) (2026-07-12)
+
+
+### Bug Fixes
+
+* resolve the webimap federation issue ([be1f1ff](https://github.com/themadorg/madmail/commit/be1f1fff82ff505a3ccd88615ecc29b84a0adebb))
+
+## [2.13.4](https://github.com/themadorg/madmail/compare/v2.13.3...v2.13.4) (2026-07-12)
+
+
+### Bug Fixes
+
+* cors issue fix for the /new also ([897a7fb](https://github.com/themadorg/madmail/commit/897a7fb47385bdb3e695af9bce604e3800f5c768))
+
+## [2.13.3](https://github.com/themadorg/madmail/compare/v2.13.2...v2.13.3) (2026-07-12)
+
+
+### Bug Fixes
+
+* **db:** add unique indexes on legacy schema ensure path ([f00fdaf](https://github.com/themadorg/madmail/commit/f00fdaf740fa0d5003c9768323d3e03bad4e23b8)), closes [#83](https://github.com/themadorg/madmail/issues/83) [#81](https://github.com/themadorg/madmail/issues/81) [#80](https://github.com/themadorg/madmail/issues/80) [#67](https://github.com/themadorg/madmail/issues/67)
+
+## [2.13.2](https://github.com/themadorg/madmail/compare/v2.13.1...v2.13.2) (2026-07-11)
+
+
+### Bug Fixes
+
+* **db:** ensure tables on Go Madmail upgrade without bare sqlx migrate ([75e9051](https://github.com/themadorg/madmail/commit/75e9051a62daf473bc8696942cf4aa6d160eaf8d)), closes [#67](https://github.com/themadorg/madmail/issues/67) [#80](https://github.com/themadorg/madmail/issues/80) [#67](https://github.com/themadorg/madmail/issues/67)
+
+## [2.13.1](https://github.com/themadorg/madmail/compare/v2.13.0...v2.13.1) (2026-07-09)
+
+
+### Bug Fixes
+
+* deliver /mxdeliv federation mail to all X-Mail-To recipients ([64c50ea](https://github.com/themadorg/madmail/commit/64c50ea586e66744565640b80d4c9afe38bb3845))
+
 # [2.13.0](https://github.com/themadorg/madmail/compare/v2.12.0...v2.13.0) (2026-07-08)